Second pass, undo larger refactor in favor of smaller MR and more code re-use
